Change page counts to size_t. Add stack lib

+++ arch/i386/include/arch.h
@@ -1,6 +1,8 @@
 #pragma once
 
-#define PAGE_SIZE 4096
+#include <stdint.h>
+
+#define PAGE_SIZE ((uintptr_t)4096)
 #define KERNEL_HIGH_VA ((uintptr_t)0xc0000000)
 
 // Address where we start allocating dynamic kernel VAs.
